pyproject.toml
[tool.poetry]
name = "recordion"
version = "0.1.0"
description = "REST API capable of managing DNS records"
authors = ["Matteo Bertucci <starchild@akarys.me>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.78.0"
asyncpg = "^0.25.0"
uvicorn = "^0.17.6"
python-jose = "^3.3.0"

[tool.poetry.dev-dependencies]
isort = "^5.10.1"
black = "^22.3.0"
flake8 = "^4.0.1"
flake8-annotations = "^2.9.0"
flake8-bugbear = "^22.4.25"
flake8-docstrings = "^1.6.0"
flake8-isort = "^4.1.1"
flake8-string-format = "^0.3.0"
flake8-todo = "^0.7"
pep8-naming = "^0.12.1"
pre-commit = "^2.19.0"
taskipy = "^1.10.2"

[tool.isort]
profile = "black"
line_length = 100

[tool.black]
line-length = 100

[tool.taskipy.tasks]
lint = { cmd = "pre-commit run --all-files", help = "Lints project files" }
precommit = { cmd = "pre-commit install", help = "Installs the pre-commit git hook" }
format = { cmd = "black --target-version py310 .", help = "Runs the black python formatter" }

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
